# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ScoutHouse API",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/v1/import/structures": {
      "post": {
        "tags": [
          "import"
        ],
        "summary": "Import Structures",
        "operationId": "import_structures_api_v1_import_structures_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "dry_run",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Dry Run"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_structures_api_v1_import_structures_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Import Structures Api V1 Import Structures Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/import/structure-open-periods": {
      "post": {
        "tags": [
          "import"
        ],
        "summary": "Import Structure Open Periods",
        "operationId": "import_structure_open_periods_api_v1_import_structure_open_periods_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "dry_run",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Dry Run"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_structure_open_periods_api_v1_import_structure_open_periods_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Import Structure Open Periods Api V1 Import Structure Open Periods Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/structures/": {
      "get": {
        "tags": [
          "structures"
        ],
        "summary": "List Structures",
        "operationId": "list_structures_api_v1_structures__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/StructureRead"
                  },
                  "type": "array",
                  "title": "Response List Structures Api V1 Structures  Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "structures"
        ],
        "summary": "Create Structure",
        "operationId": "create_structure_api_v1_structures__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StructureCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StructureRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_import_structure_open_periods_api_v1_import_structure_open_periods_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_import_structure_open_periods_api_v1_import_structure_open_periods_post"
      },
      "Body_import_structures_api_v1_import_structures_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_import_structures_api_v1_import_structures_post"
      },
...
      "CostBand": {
        "type": "string",
        "enum": [
          "cheap",
          "medium",
          "expensive"
        ],
        "title": "CostBand"
      },
...
      "FirePolicy": {
        "type": "string",
        "enum": [
          "allowed",
          "with_permit",
          "forbidden"
        ],
        "title": "FirePolicy"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "StructureAvailabilityRead": {
        "properties": {
          "season": {
            "$ref": "#/components/schemas/StructureSeason"
          },
          "units": {
            "items": {
              "$ref": "#/components/schemas/StructureUnit"
            },
            "type": "array",
            "title": "Units"
          },
          "capacity_min": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Capacity Min"
          },
          "capacity_max": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Capacity Max"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "season",
          "id"
        ],
        "title": "StructureAvailabilityRead"
      },
...
      "StructureCostOptionRead": {
        "properties": {
          "model": {
            "$ref": "#/components/schemas/StructureCostModel"
          },
          "amount": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Amount"
          },
          "currency": {
            "type": "string",
            "maxLength": 3,
            "minLength": 3,
            "title": "Currency",
            "default": "EUR"
          },
          "deposit": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deposit"
          },
          "city_tax_per_night": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "City Tax Per Night"
          },
          "utilities_flat": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Utilities Flat"
          },
          "age_rules": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Age Rules"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "model",
          "amount",
          "id"
        ],
        "title": "StructureCostOptionRead"
      },
...
      "StructureRead": {
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "title": "Name"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "province": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2
              },
              {
                "type": "null"
              }
            ],
            "title": "Province"
          },
          "address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "latitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "longitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          },
          "altitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Altitude"
          },
          "type": {
            "$ref": "#/components/schemas/StructureType"
          },
          "indoor_beds": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Indoor Beds"
          },
          "indoor_bathrooms": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Indoor Bathrooms"
          },
          "indoor_showers": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Indoor Showers"
          },
          "indoor_activity_rooms": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Indoor Activity Rooms"
          },
          "has_kitchen": {
            "type": "boolean",
            "title": "Has Kitchen",
            "default": false
          },
          "hot_water": {
            "type": "boolean",
            "title": "Hot Water",
            "default": false
          },
          "land_area_m2": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Land Area M2"
          },
          "shelter_on_field": {
            "type": "boolean",
            "title": "Shelter On Field",
            "default": false
          },
          "water_sources": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/WaterSource"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Water Sources"
          },
          "electricity_available": {
            "type": "boolean",
            "title": "Electricity Available",
            "default": false
          },
          "fire_policy": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FirePolicy"
              },
              {
                "type": "null"
              }
            ]
          },
          "access_by_car": {
            "type": "boolean",
            "title": "Access By Car",
            "default": false
          },
          "access_by_coach": {
            "type": "boolean",
            "title": "Access By Coach",
            "default": false
          },
          "access_by_public_transport": {
            "type": "boolean",
            "title": "Access By Public Transport",
            "default": false
          },
          "coach_turning_area": {
            "type": "boolean",
            "title": "Coach Turning Area",
            "default": false
          },
          "nearest_bus_stop": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Nearest Bus Stop"
          },
          "weekend_only": {
            "type": "boolean",
            "title": "Weekend Only",
            "default": false
          },
          "has_field_poles": {
            "type": "boolean",
            "title": "Has Field Poles",
            "default": false
          },
          "pit_latrine_allowed": {
            "type": "boolean",
            "title": "Pit Latrine Allowed",
            "default": false
          },
          "website_urls": {
            "items": {
              "type": "string",
              "minLength": 1,
              "format": "uri"
            },
            "type": "array",
            "title": "Website Urls"
          },
          "notes_logistics": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes Logistics"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "estimated_cost": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Cost"
          },
          "cost_band": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CostBand"
              },
              {
                "type": "null"
              }
            ]
          },
          "availabilities": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/StructureAvailabilityRead"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Availabilities"
          },
          "cost_options": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/StructureCostOptionRead"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost Options"
          },
          "contacts": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ContactRead"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contacts"
          },
          "open_periods": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/StructureOpenPeriodRead"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Open Periods"
          }
        },
        "type": "object",
        "required": [
          "name",
          "slug",
          "type",
          "id",
          "created_at"
        ],
        "title": "StructureRead"
      },
...
      "StructureSeason": {
        "type": "string",
        "enum": [
          "winter",
          "spring",
          "summer",
          "autumn"
        ],
        "title": "StructureSeason"
      },
      "StructureType": {
        "type": "string",
        "enum": [
          "house",
          "land",
          "mixed"
        ],
        "title": "StructureType"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WaterSource": {
        "type": "string",
        "enum": [
          "none",
          "fountain",
          "tap",
          "river"
        ],
        "title": "WaterSource"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}